+++ CHANGELOG.md
@@ -45,6 +45,7 @@
  * STORM-626: Add script to print out the merge command for a given pull request.
  * STORM-601: Make jira-github-join ignore case.
  * STORM-667: Incorrect capitalization "SHell" in Multilang-protocol.md
+ * STORM-637: Integrate PartialKeyGrouping into storm API
 
 ## 0.9.3-rc2
  * STORM-558: change "swap!" to "reset!" to fix assignment-versions in supervisor